requirements.txt
Flask==2.0.1 
google-cloud-vision==3.1.0 
google-auth==2.9.1 
unidecode==1.3.4 
nltk==3.7 
Pillow==9.2.0 
pytesseract==0.3.9 
numpy==1.23.1 
opencv-python==4.6.0.66